Fix async commit creation bug
sometimes when trying to to create a commit, the repository id was not loaded in database. Since the repository_id its mandatory to create a commit, a new rule in the serializer was created to convert the repository_id to repository_name and vice-versa. This solves the problem of commit's repository name not being loaded in frontend as the commit creation error in POST requests